refactor: centralize PersonaProfile
